Added images of text and clean up
This is a pull request on Steve branch which is intended to clean up the accessible name SC. My fear is that although it was previously not consistent with some of the terminology in other SC (Accessible Name, Label) it actually was fairly easy to understand. I'm trying to keep Steve's clean up of terms but restore some of its previously user friendly language which made it clear that the string of text from the visible label is also presented in the invisible name. I fear that if readers have to go digging through the definitions to understand that a label is visible and a name may not be, will cause some to scratch their heads. This was a major complaint against 2.0